Refactor: Organize backend API and analytics modules
Consolidate API routes into backend/api/v1 and move analytics logic to services.
