+++ backend/api/v1/__init__.py
@@ -1,0 +1,7 @@
+"""API v1 routes - versioned API structure."""
+
+from backend.api.v1.billing import billing_router
+from backend.api.v1.workflows import workflow_router
+from backend.api.v1.workflow_automation import router as workflow_automation_router
+
+__all__ = ["billing_router", "workflow_router", "workflow_automation_router"]
